requirements.txt
# Архивы
patool==1.12

# PDF — основной парсер (заменяет PyPDF2)
pymupdf==1.24.5
# PyPDF2==3.0.0  # оставлен как резерв, можно раскомментировать

# OCR
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0      # конвертация страниц PDF в картинки для OCR

# DOCX
python-docx==1.1.0

# FB2 и EPUB (XML)
lxml==4.9.3

# DJVU — Python-биндинг djvu недоступен на PyPI для Python 3.x.
# DjVu-файлы обрабатываются через внешний инструмент ddjvu (из DjVuLibre),
# который должен быть установлен отдельно и прописан в PATH.
# Скачать: https://sourceforge.net/projects/djvu/files/DjVuLibre_Windows/

# HTTP клиент для Ollama (входит в stdlib, но явно фиксируем версию)
requests==2.32.3


# MOBI/AZW (опционально — нужен только если EXTH метаданных недостаточно)
mobi==0.4.1
